cdk/requirements.txt
aws-cdk-lib>=2.124.0
constructs>=10.0.0
aws-cdk.lambda-layer-kubectl-v32>=2.0.0
python-dotenv>=1.0.0